Only provision a new lambda if it isn't there already
That piece of code is a bit messy because the API seems to be a bit
silly. With all the other requests, we received a decent response that,
in the error case, contained a :Code and a :Message. For this one, it's
simply a message, in some responses with a capital M and in some with a
small m.

The code is not nice and I don't like it, but it does the job for now
and I don't wanna overengineer it, as long as it works.