Implement simple deleter without respecting dependency breaks
A resource accepts an additional deleter function. It will be called
when a resource is in :delete state. The function can be used to clean
up any real-world resources, its result value is the new resource. The
resource state automatically switches to :deleted.

We don't remove deleted resources from state because that would cause
their immediate recreation (as long as they still are in the brood).
Immediate recreation can effectively prevent deleting multiple resources
if their dependencies are in the wrong order: the resource that still
depends would never be deleted, the dependee would just flicker out of
and into existance.

Respecting dependencies when deleting will follow in a later patch.