Implement after function for resources
Sometimes you want to create a resource and if successful do some
additional configuration that does not require a resource in its own
right. For example you might want to create a AWS Lambda and if
successful add permissions to it. Those permissions need not be deleted
manually: if you delete the lambda, the permissions are gone too.

Doing that kind of configuration inside the breeder destroys its
atomicity. If the additional configuration fails, the resource is
created but not reflected in the state.

After functions can be used to perfom that kind of config. If the after
function fails, the resource will appear in the state, but it will be in
:failed state. You can easily recover by deleting and recreating
resources.